+++ R/misc.R
@@ -13,7 +13,7 @@ NULL
 #' @seealso [base::options()].
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' # check if Gurobi is installed
 #' is.GurobiInstalled()
 #'
@@ -66,82 +66,6 @@ is.GurobiInstalled <- function(verbose = TRUE) {
   if (!pkgs.installed[[1]])
     return(FALSE)
   return(TRUE)
-}
-
-#' Test if GDAL is installed on computer
-#'
-#' This function tests if GDAL is installed on the computer.
-#' If not, download it here: <http://download.osgeo.org/gdal>.
-#'
-#' @return `logical` is GDAL installed?
-#'
-#' @seealso [gdalUtils::gdal_setInstallation()].
-#'
-#' @examples
-#' # check if gdal is installed on system
-#' \donttest{
-#' is.gdalInstalled()
-#' }
-#'
-#' @export
-is.gdalInstalled <- function() {
-  suppressWarnings(findGdalInstallationPaths())
-  return(!is.null(getOption("gdalUtils_gdalPath")))
-}
-
-#' Rasterize polygon data using GDAL
-#'
-#' This function converts a [sp::SpatialPolygonsDataFrame()] to a
-#' [raster::raster()] using GDAL. It is expected to be faster than
-#' [raster::rasterize()] for large datasets. However, it will be
-#' significantly slower for small datasets because the data will need to be
-#' written and read from disk.
-#'
-#' @param x [sp::SpatialPolygonsDataFrame()] object.
-#'
-#' @param y [raster::raster()] with dimensions, extent, and
-#'   resolution to be used as a template for new raster.
-#'
-#' @param field `character` column name with values to burn into the
-#'   output raster. If not supplied, default behaviour is to burn polygon
-#'   indices into the [raster::raster()].
-#'
-#' @return [raster::raster()] object.
-#'
-#' @seealso [raster::rasterize()], [is.gdalInstalled()].
-#'
-#' @examples
-#' \donttest{
-#' # load dat
-#' data(cs_pus,cs_spp)
-#'
-#' # rasterize spatial polygon data
-#' x <- rasterizeGDAL(cs_pus[1:5,], cs_spp[[1]])
-#'
-#' # plot data
-#' par(mfrow = c(1,2))
-#' plot(cs_pus[1:5,], main = "original data")
-#' plot(x, main = "rasterized data")
-#' }
-#'
-#' @export
-rasterizeGDAL <- function(x, y, field = NULL) {
-  assertthat::assert_that(inherits(x, "SpatialPolygonsDataFrame"),
-                          inherits(y, "RasterLayer"),
-                          is.null(field) || (assertthat::is.string(field) &&
-                                             field %in% names(x@data)))
-  if (is.null(field)) {
-    x@data$id <- seq_len(nrow(x@data))
-    field <- "id"
-  }
-  rgdal::writeOGR(x, tempdir(), "polys", driver = "ESRI Shapefile",
-                  overwrite_layer = TRUE)
-  raster::writeRaster(raster::setValues(y, NA), file.path(tempdir(),
-                      "rast.tif"), NAflag = -9999, overwrite = TRUE)
-  return(gdalUtils::gdal_rasterize(file.path(tempdir(), "polys.shp"),
-                                  file.path(tempdir(), "rast.tif"),
-                                  l = "polys", a = field,
-                                  output_Raster = TRUE)[[1]])
 }
 
 #' Blank raster