+++ R/sim.pus.R
@@ -31,7 +31,7 @@ NULL
 #'
 #' # generate 225 rectangular pus arranged in a square
 #' y <- sim.pus(225, xmn = -5, xmx = 10, ymn = -5, ymx = 5)
-#' \donttest{
+#' \dontrun{
 #' par(mfrow = c(1, 2))
 #' plot(x, main = "x")
 #' plot(y, main = "y")